requirements.txt
aiosqlite==0.17.0
annotated-types==0.5.0
anyio==3.7.1
asyncpg==0.28.0
click==8.1.7
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.103.1
gunicorn==21.2.0
h11==0.14.0
httptools==0.6.0
idna==3.4
iso8601==1.1.0
packaging==23.1
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
pytz==2023.3.post1
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tortoise-orm==0.20.0
typing_extensions==4.8.0
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
